+++ mobile/android/android-components/components/feature/session/src/test/java/mozilla/components/feature/session/HistoryDelegateTest.kt
@@ -97,6 +97,10 @@ class HistoryDelegateTest {
                 fail()
             }
 
+            override suspend fun deleteVisit(url: String, timestamp: Long) {
+                fail()
+            }
+
             override suspend fun prune() {
                 fail()
             }